refactor: Separating responsibilities.
