+++ adventures.and.beyond/src/helper/Constant.java
@@ -23,63 +23,4 @@ public class Constant {
     public static final int FRAME_RATE_PER_SECOND = 60;//fps=60
 
     public  static final int PLAYER_UP_SCALE=2;
-
-    public static BufferedImage getImage(String imgPath) {
-        try {
-            InputStream resource = Constant.class.getResourceAsStream(imgPath);
-            if (resource == null) {
-                throw new IOException("Custom error:Image resource is null.Given path: " + imgPath);
-            }
-
-            return ImageIO.read(resource);
-        } catch (IOException exception) {
-            exception.printStackTrace();
-        }
-        return null;
-    }
-
-    public static BufferedReader readTextFile(String filePath){
-        try{
-            InputStream resource = Constant.class.getResourceAsStream(filePath);
-
-            if (resource == null) {
-                throw new IOException("Custom error:Image resource is null.Given path: " + filePath);
-            }
-
-            return new BufferedReader(new InputStreamReader(resource));
-        }catch (Exception exception){
-            System.out.println("Custom Error: Error occurred when loading the map. Given File path: "+filePath);
-            exception.fillInStackTrace();
-        }
-        return null;
-    }
-
-    public static  int[][] getMapMatrix(String filePath){
-        BufferedReader reader = readTextFile(filePath);
-        assert reader != null;
-        int totalRows = Constant.MAX_SCREEN_ROW;
-        int totalColumns= Constant.MAX_SCREEN_COLUMN;
-
-        int[][] matrix = new int[totalRows][totalColumns];
-        try {
-
-            for (int row = 0; row < totalRows; row++) {
-
-                String line = reader.readLine();
-                String[] numericStr = line.split("");
-                System.out.println("");
-                for (int col = 0; col < totalColumns; col++) {
-                    int value = Integer.parseInt(numericStr[col]);
-                    System.out.print(value);
-                    matrix[row][col]=value;
-                }
-
-            }
-        }catch (IOException exception){
-            System.out.println("Custom Error:Error occurred when reading the map line by line.FilePath: "+filePath);
-            exception.fillInStackTrace();
-        }
-        return matrix;
-    }
-
 }
